# src/main/resources/openapi/voetballer-app-rest.yaml
openapi: 3.0.3
info:
  title: Voetballer API
  version: 1.0.0
  description: API voor het beheren van voetballers en het verzenden ervan naar een ActiveMQ-queue.

servers:
  - url: http://localhost:8080

paths:
  /voetballers:
    get:
      operationId: getVoetballers
      summary: Haal alle voetballers op
      responses:
        '200':
          description: Lijst van voetballers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Voetballer'

    post:
      operationId: createVoetballer
      summary: Maak een nieuwe voetballer aan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoetballerRequest'
      responses:
        '201':
          description: Voetballer aangemaakt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voetballer'

  /voetballers/{id}:
    get:
      operationId: getVoetballerById
      summary: Haal een voetballer op via ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Voetballer gevonden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voetballer'
        '404':
          description: Niet gevonden

    put:
      operationId: updateVoetballer
      summary: Wijzig een bestaande voetballer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoetballerRequest'
      responses:
        '200':
          description: Voetballer gewijzigd
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voetballer'
        '404':
          description: Niet gevonden

    delete:
      operationId: deleteVoetballer
      summary: Verwijder een voetballer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Verwijderd
        '404':
          description: Niet gevonden

  /voetballers/{id}/send:
    post:
      operationId: sendVoetballer
      summary: Stuur een voetballer naar de ActiveMQ-queue
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Voetballer naar queue gestuurd
        '404':
          description: Niet gevonden

components:
  schemas:
    Voetballer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        naam:
          type: string
          maxLength: 100
          example: Stan de Voetballer
        positie:
          type: string
          example: Aanvaller
        team:
          type: string
          example: SV de Weide
      required:
        - naam
        - positie
        - team

    VoetballerRequest:
      type: object
      properties:
        naam:
          type: string
          maxLength: 100
          example: Marco Jansen
        positie:
          type: string
          example: Verdediger
        team:
          type: string
          example: SV de Weide
      required:
        - naam
        - positie
        - team